requirements.txt
fortnite-api
pandas
numpy
streamlit
python-dotenv